refactor: Adjsut file name to align with the page name
